+++ tests/functional/translate/cram/translate-with-gff-and-gene.t
@@ -18,8 +18,8 @@ Translate amino acids for genes using a GFF3 file where the gene names are store
   >   --ancestral-sequences "${DATA}/zika/nt_muts.json" \
   >   --reference-sequence genemap.gff \
   >   --output-node-data aa_muts.json
-  Validating schema of '.+/nt_muts.json'... (re)
   Read in 2 features from reference sequence file
+  Validating schema of '.+/nt_muts.json'... (re)
   amino acid mutations written to .* (re)
 
   $ python3 "${SCRIPTS}/diff_jsons.py" \